+++ karbide-core/src/commonTest/kotlin/dev/karmakrafts/karbide/BitSinkTest.kt
@@ -9,13 +9,13 @@ import kotlinx.io.readUShort
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class BitWriterTest {
+class BitSinkTest {
     @Test
     fun `Write bits within same byte`() {
         val buffer = Buffer()
-        buffer.bitWriter().use { writer ->
-            writer.writeNibble(0b0000U)
-            writer.writeNibble(0b0011U)
+        buffer.bitSink().use { sink ->
+            sink.writeNibble(0b0000U)
+            sink.writeNibble(0b0011U)
         }
         assertEquals(3U.toUByte(), buffer.readUByte())
     }
@@ -23,10 +23,10 @@ class BitWriterTest {
     @Test
     fun `Write bits beyond byte boundary`() {
         val buffer = Buffer()
-        buffer.bitWriter().use { writer ->
-            writer.writeUByte(0U)
-            writer.writeNibble(0b0000U)
-            writer.writeNibble(0b0011U)
+        buffer.bitSink().use { sink ->
+            sink.writeUByte(0U)
+            sink.writeNibble(0b0000U)
+            sink.writeNibble(0b0011U)
         }
         buffer.skip(UByte.SIZE_BYTES.toLong())
         assertEquals(3U.toUByte(), buffer.readUByte())
@@ -35,18 +35,18 @@ class BitWriterTest {
     @Test
     fun `Write numeric types`() {
         val buffer = Buffer()
-        buffer.bitWriter().use { writer ->
-            writer.writeByte(0x12)
-            writer.writeShort(0x1234)
-            writer.writeShortLe(0x1234)
-            writer.writeInt(0x12345678)
-            writer.writeIntLe(0x12345678)
-            writer.writeLong(0x1234567890ABCDEFL)
-            writer.writeLongLe(0x1234567890ABCDEFL)
-            writer.writeFloat(1.234f)
-            writer.writeFloatLe(1.234f)
-            writer.writeDouble(1.23456789)
-            writer.writeDoubleLe(1.23456789)
+        buffer.bitSink().use { sink ->
+            sink.writeByte(0x12)
+            sink.writeShort(0x1234)
+            sink.writeShortLe(0x1234)
+            sink.writeInt(0x12345678)
+            sink.writeIntLe(0x12345678)
+            sink.writeLong(0x1234567890ABCDEFL)
+            sink.writeLongLe(0x1234567890ABCDEFL)
+            sink.writeFloat(1.234f)
+            sink.writeFloatLe(1.234f)
+            sink.writeDouble(1.23456789)
+            sink.writeDoubleLe(1.23456789)
         }
 
         assertEquals(0x12.toByte(), buffer.readByte())
@@ -65,14 +65,14 @@ class BitWriterTest {
     @Test
     fun `Write unsigned numeric types`() {
         val buffer = Buffer()
-        buffer.bitWriter().use { writer ->
-            writer.writeUByte(0x12U)
-            writer.writeUShort(0x1234U)
-            writer.writeUShortLe(0x1234U)
-            writer.writeUInt(0x12345678U)
-            writer.writeUIntLe(0x12345678U)
-            writer.writeULong(0x1234567890ABCDEFUL)
-            writer.writeULongLe(0x1234567890ABCDEFUL)
+        buffer.bitSink().use { sink ->
+            sink.writeUByte(0x12U)
+            sink.writeUShort(0x1234U)
+            sink.writeUShortLe(0x1234U)
+            sink.writeUInt(0x12345678U)
+            sink.writeUIntLe(0x12345678U)
+            sink.writeULong(0x1234567890ABCDEFUL)
+            sink.writeULongLe(0x1234567890ABCDEFUL)
         }
 
         assertEquals(0x12U.toUByte(), buffer.readUByte())
@@ -88,8 +88,8 @@ class BitWriterTest {
     fun `Write byte array`() {
         val buffer = Buffer()
         val data = byteArrayOf(0x01, 0x02, 0x03, 0x04, 0x05)
-        buffer.bitWriter().use { writer ->
-            writer.writeByteArray(data)
+        buffer.bitSink().use { sink ->
+            sink.writeByteArray(data)
         }
         val readData = buffer.readByteArray(data.size)
         for (i in data.indices) {
@@ -100,13 +100,13 @@ class BitWriterTest {
     @Test
     fun `Write padding`() {
         val buffer = Buffer()
-        buffer.bitWriter().use { writer ->
-            writer.writeBit(1U)
-            writer.padBit(0U)
-            writer.padNibble(1U)
-            writer.padNibbles(2, 0U)
-            writer.padByte(1U)
-            writer.padBytes(2, 0U)
+        buffer.bitSink().use { sink ->
+            sink.writeBit(1U)
+            sink.padBit(0U)
+            sink.padNibble(1U)
+            sink.padNibbles(2, 0U)
+            sink.padByte(1U)
+            sink.padBytes(2, 0U)
         }
 
         assertEquals(0xBC.toUByte(), buffer.readUByte())