+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ### Added
 
-- `BitReader` interface for reading bit streams
-- `BitWriter` interface for writing bit streams
+- `BitSource` interface for reading bit streams
+- `BitSink` interface for writing bit streams
 - Highly optimized `reverseBytes` extensions for all integral types for swapping endianess